backend/requirements.txt
# FastAPI and API dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic>=2.9.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.36
alembic==1.13.1
aiosqlite==0.19.0

# GraphQL
strawberry-graphql[fastapi]>=0.250.0
graphql-core>=3.2.3

# AI and ML
langchain>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
faiss-cpu==1.13.0
sentence-transformers==2.3.1
numpy>=1.26.4

# Ollama
ollama==0.1.6

# HTTP clients
httpx>=0.25.2,<0.26.0
aiohttp==3.9.1
requests==2.31.0

# Data processing
pandas>=2.2.0
python-dateutil==2.8.2

# Environment and Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx>=0.25.2,<0.26.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# CORS
fastapi-cors==0.0.6

# PDF generation
reportlab==4.0.9
matplotlib>=3.9.0
Pillow>=10.3.0

# Logging
loguru==0.7.2

# Validation
email-validator==2.1.0
